refactor: split GitHub embed plugin
